[HOTSPOT-81] refactor: 가족 목록, 가족 디테일, 정책 관리 리팩토링
