+++ apps/admin/src/domains/family/index.ts
@@ -1,0 +1,9 @@
+export type {
+  FamilyDetail,
+  FamilyItem,
+  FamilyList,
+  FamilyListRequest,
+  FamilyPolicy,
+} from './model/types';
+export { useFamilyDetail } from './model/useFamilyDetail';
+export { useFamilyList } from './model/useFamilyList';
