+++ apps/admin/src/domains/member-control/index.ts
@@ -1,0 +1,2 @@
+export type { MemberControl, MemberControlItem } from './model/types';
+export { useFamilyDetailControl } from './model/useFamilyDetailControl';
