+++ apps/admin/src/domains/member-policy/api/getFamilyDetailPolicyClient.ts
@@ -1,6 +1,6 @@
+import type { FamilyPolicy } from '@/domains/family';
 import { api } from '@/shared/api/client';
 import type { ApiResponse } from '@/shared/api/types';
-import type { FamilyPolicy } from '../../../families/model/types';
 
 export const getFamilyDetailPolicyClient = async (familyId: number) => {
   const { data } = await api.get<ApiResponse<FamilyPolicy[]>>(
