+++ apps/admin/src/domains/policy/index.ts
@@ -1,0 +1,1 @@
+export { useBlocked, usePolicy } from './model/usePolicy';
