--- apps/admin/src/features/families/index.ts
@@ -1,4 +1,0 @@
-export { FamilyDetailControlTab } from '../family-control/ui/FamilyDetailControlTab';
-export { FamilyDetailPolicyTab } from '../family-policy/member-info/ui/FamilyDetailPolicyTab';
-export { useFamilyDetail } from './model/useFamilyDetail';
-export { useFamilyList } from './model/useFamilyList';
