+++ apps/admin/src/features/family-control/index.ts
@@ -1,0 +1,1 @@
+export { FamilyDetailControlTab } from './ui/FamilyDetailControlTab';
