+++ apps/admin/src/features/family-policy/member-info/index.ts
@@ -1,0 +1,1 @@
+export { FamilyDetailPolicyTab } from './ui/FamilyDetailPolicyTab';
