+++ apps/admin/src/shared/ui/index.ts
@@ -5,3 +5,4 @@ export { SearchBar } from './search-bar/SearchBar';
 export { SideBar } from './side-bar/SideBar';
 export type { Column } from './table/Table';
 export { Table } from './table/Table';
+export { UserProfileIcon } from './user-profile-icon/UserProfileIcon';
